+++ examples/resources/coolify_docker_image_application/import.sh
@@ -1,0 +1,1 @@
+terraform import coolify_docker_image_application.nginx <application-uuid>